Check for nullability assertions on bodies in CodegenAnnotatingVisitor
RuntimeAssertionsOnDeclarationBodyChecker is not really a declaration
checker, this caused some potentially problematic code
(such as expecting BindingTrace instead of an arbitrary DiagnosticSink).

It makes sense to run this check independently in
CodegenAnnotatingVisitor.

Note that this makes sense only when actual code for bodies is
generated.
Otherwise (e.g., if it's invoked from light classes generation)
it might cause recursion in type resolution.